Created DBAdder Queries
- Created the DBAdder addNewAccountQuery.
- Had to do some modifications to the Account class and the
AccountManager. The accountID was set as a string when it should
be an integer.  Also added another constructor to the Account class
as there was no constructor to input all fields without a SubAccount.
- Continued to try to get the addNewPriceTier query to work but kept
getting the same error.  I thought it was because of the double type
but when I created the Account query there was also a double type for
the balance and I had no issues.  I ran the prepared statement in
phpmyadmin and it worked with no error.  Still unsure what the issue
could be.